Move ConstraintType and ConstraintMap classes to kinematic_constraints.py
