+++ src/jaxsim/rbda/__init__.py
@@ -1,4 +1,11 @@
+# ruff: isort: skip_file
 from . import contacts
+from .kinematic_constraints import (
+    compute_constraint_baumgarte_term,
+    compute_constraint_jacobians,
+    compute_constraint_jacobians_derivative,
+    compute_constraint_transforms,
+)
 from .aba import aba
 from .collidable_points import collidable_points_pos_vel
 from .crba import crba
@@ -7,11 +14,5 @@ from .jacobian import (
     jacobian,
     jacobian_derivative_full_doubly_left,
     jacobian_full_doubly_left,
-)
-from .kinematic_constraints import (
-    compute_constraint_baumgarte_term,
-    compute_constraint_jacobians,
-    compute_constraint_jacobians_derivative,
-    compute_constraint_transforms,
 )
 from .rnea import rnea
